web/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
cffi==1.15.0
click==8.1.3
Flask==2.1.2
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
protobuf==4.21.2
pycparser==2.21
python-dotenv==0.20.0
Werkzeug==2.1.2
WTForms==3.0.1
